Added toggleMoreActions to Markdown Preview only. When it is in editor mode, only editor display ToggleMoreActions.
