+++ src/sql/parts/notebook/cellViews/code.component.html
@@ -9,6 +9,6 @@
 	</div>
 	<div #editor class="editor" style="flex: 1 1 auto; overflow: hidden;">
 	</div>
-	<div #moreactions class="toolbar" style="flex: 0 0 auto; display: flex; flex-flow:column; width: 20px; min-height: 20px; max-height: 20px; padding-top: 0px; orientation: portrait">
+	<div #moreactions class="moreActions" style="flex: 0 0 auto; display: flex; flex-flow:column;">
 	</div>
 </div>
