+++ src/sql/parts/notebook/cellViews/code.css
@@ -45,10 +45,6 @@ code-component .carbon-taskbar .icon {
 	width: 40px;
 }
 
-code-component .action-label.icon.toggle-more {
-	height: 20px;
-	width: 20px;
-}
 
 code-component .carbon-taskbar.monaco-toolbar .monaco-action-bar.animated .actions-container
 {
