+++ src/sql/parts/notebook/notebook.css
@@ -75,3 +75,12 @@
 .hc-black .notebookEditor .notebook-button.icon-save{
 	background-image: url("./media/dark/save_inverse.svg");
 }
+
+.moreActions .action-label.icon.toggle-more {
+	height: 20px;
+	width: 20px;
+	min-height: 20px;
+	max-height: 20px;
+	padding-top: 0px;
+	orientation: portrait;
+}
